Move some directories from common/ to core/ to prevent circular dependencies
