+++ packages/core/src/common/index.ts
@@ -1,8 +1,6 @@
 export * from './auth';
 export * from './encoding';
 export * from './file';
-export * from './logging';
-export * from './templates';
 export * from './tokens';
 export * from './utils';
 export * from './validation';
