+++ packages/core/src/core/index.ts
@@ -2,8 +2,10 @@ export * from './app.controller.interface';
 export * from './class.interface';
 export { createController } from './controllers';
 export * from './http';
+export * from './logging';
 export * from './openapi';
 export * from './hooks';
 export * from './routes';
+export * from './templates';
 export * from './config';
 export { createService, dependency, Dependency, ServiceManager } from './service-manager';
