+++ packages/core/src/core/templates/500.debug.html
