+++ packages/core/src/core/templates/index.ts
