+++ packages/core/src/index.ts
@@ -36,10 +36,6 @@ export {
   hashPassword,
   passwordHashNeedsToBeRefreshed,
   isInFile,
-  Logger,
-  render,
-  renderToString,
-  renderError,
   signToken,
   streamToBuffer,
   verifyPassword,
@@ -214,6 +210,11 @@ export {
   isHttpResponseSuccess,
   isHttpResponseTooManyRequests,
   isHttpResponseUnauthorized,
+
+  Logger,
+  render,
+  renderToString,
+  renderError,
 } from './core';
 export {
   OPENAPI_SERVICE_ID,
